Refactor and clean implementation
Add SmartShipClient that handles the communication with the API. Drop all environment variable based configuration. Now a client is set up separately with the authorization tokens and then the client `.send_shipment(shipment)` method is used to send a shipment. The `Shipment` class still has a helper `.send(authorization)` method. It must be given the authorization tokens if it is used.

Clean up and add tests for the Posti shipment creator wrapper. Add more tests for the rests of the functionality that exists.

Add coverage to dev requirements.