setup.py: Ensure recent enough setuptools version
Old setuptools versions doesn't support defining the metadata in
setup.cfg.  Make sure that the version is recent enough by adding
setup_requires.

Also add the if-main guard to the setup.py script so that no tool will
run the setup script even if the tool is stupid enough to import it.